requirements.txt
tensorflow-gpu==1.13.1
pillow==8.1.2
lxml==4.6.3
Cython==0.29.22
contextlib2
jupyter==1.0.0
matplotlib==3.3.4
pandas==1.1.5
opencv-python==4.5.1.48
streamlit==0.79.0
numpy==1.19.5
